backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.0
httpx==0.27.2
openai==1.54.0
ics==0.7.2
python-multipart==0.0.9
pillow==10.4.0